requirements.txt
requests
# Используем psycopg2-binary вместо psycopg2 для значительного ускорения
# деплоя на Railway, т.к. эта версия не требует компиляции.
psycopg2-binary
vk_api
Flask
openpyxl
gunicorn
APScheduler
google-cloud-aiplatform
pytz
google-oauth2-tool
google-generativeai
python-dotenv
